Replaced Open Hub script by generated code
The script uses document.write() and therefore cannot be loaded
asynchronously. We don't want it to load it synchronously either,
since that would delay the displaying of the main content.

I reported the issue on their forum:
  https://www.openhub.net/forums/11/topics/9793

In the mean time, I put the <iframe> tag generated by the script
in the page. This is not future proof, but it works for now.
Unfortunately that tag uses attributes that are obsolete in HTML5,
but if I try to replace those attributes by equivalent CSS the page
mis-renders in Firefox (and possibly other browsers I didn't try).